requirements.txt
python-dotenv
praw
streamlit

groq